[KLDV-3091] 오픈링크 개설 API (6/N)

+++ src/main/java/com/klipwallet/membership/repository/ChatRoomNftRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.klipwallet.membership.entity.ChatRoomNft;
 
-public interface ChatRoomNftRepository extends JpaRepository<ChatRoomNft, Integer> {
+public interface ChatRoomNftRepository extends JpaRepository<ChatRoomNft, Long> {
 }
